backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
apscheduler
torch
numpy
pandas
scikit-learn
websockets
